refactor: move animated-scroll-to to scrollbar-helper
